# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9044",
    "basePath": "/",
    "paths": {
...
        "/oidc/v1/auth": {
            "get": {
                "description": "get the home page.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "root"
                ],
                "summary": "get the home page.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "client_id requested",
                        "name": "client_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "response_type requested",
                        "name": "response_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "\"openid profile email\"",
                        "description": "scope requested",
                        "name": "scope",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "state requested",
                        "name": "state",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "redirect_uri requested",
                        "name": "redirect_uri",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "audience requested",
                        "name": "audience",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "PKCE challenge code",
                        "name": "code_challenge",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"S256\"",
                        "description": "PKCE challenge method",
                        "name": "code_challenge_method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "acr_values requested",
                        "name": "acr_values",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}